+++ ux/UX-Employees-Swap-Name.js
@@ -1,0 +1,1 @@
+scriptInject('ux/src/UX-Employees-Swap-Name-Script.js')