+++ ux/UX-Role-History-Buttons.js
@@ -7,10 +7,16 @@ window.addEventListener('message',function(event){
         case "copyRoles":
             copyRoles()
             break;
+        case "uxRolesFilter":
+            filterUX()
+            break;
+        case "classicRolesFilter":
+            filterClassic()
+            break;
     }
 })
 
-function copyRoles(){
+function copyRoles(){ //
     var a = [];
     var status;
     $('tbody > tr').each(
@@ -23,4 +29,34 @@ function copyRoles(){
     chrome.storage.local.set({roles:a},function(){
             console.log('Value is set to '+ a);
         });
+    var u = $('tbody > tr > td')[0].children[0].textContent
+    chrome.storage.local.set({copied:u},function(){
+            console.log('Value copied from '+ u);
+        });
     }
+function filterUX(){
+    var role;
+    $('tbody > tr').each(
+        function(){
+            role = $(this).find('td').eq(2).text();
+            if (!role.startsWith('UX')){
+                this.style = 'display: none;'
+            }
+            else{
+                this.removeAttribute('style')
+            }
+        })
+}
+function filterClassic(){
+    var role;
+    $('tbody > tr').each(
+        function(){
+            role = $(this).find('td').eq(2).text();
+            if (role.startsWith('UX')){
+                this.style = 'display: none;'
+            }
+            else{
+                this.removeAttribute('style')
+            }
+        })
+}