+++ ux/src/UX-Position-Role-Association-Buttons-Script.js
@@ -1,6 +1,11 @@
+const dataset = document.currentScript.getAttribute('data-params')
 function createTheButtons(){
+    console.log('script.js - dataset: ', dataset)
+    const params = JSON.parse(dataset);
+    console.log('script.js - createTheButtons function - injected params: ', params);
+    let pasteText = params.pasted
         uxCreateButton('copyRoles','Copy Roles',null,null,false,'Roles Copied');
-        uxCreateButton('pasteRoles','Paste Roles',null,null,false,'Roles Pasted');
+        uxCreateButton('pasteRoles',`Paste Roles (${pasteText})`,null,null,false,'Roles Pasted');
         document.getElementById ("copyRoles").addEventListener (
             "mouseup", buttonCallback, false
             );